requirements.txt
paho-mqtt
dht11
RPi.GPIO
tox
PyYAML
temporenc
requests
python-switchbot